refactor: rename directory, clean up files (no logic change)
