+++ src/app/components/map/layers/particles/config.ts
