+++ src/app/components/map/layers/particles/trips/evenly-spaced-points.ts
