+++ src/app/components/map/layers/particles/trips/particle-generator.ts
