+++ src/app/components/map/layers/particles/trips/particle-scaling.ts
