+++ src/app/components/map/layers/particles/trips/point-spacing.ts
