+++ src/app/components/map/layers/particles/trips/trip-path.ts
