+++ src/app/components/map/layers/particles/types.ts
