+++ src/app/components/map/state/machine.ts
@@ -510,8 +510,6 @@ export const globeViewMachine = createMachine(
           "visible"
         );
         m.setLayoutProperty("destination-areas-fill", "visibility", "visible");
-        m.setLayoutProperty("area-flows", "visibility", "visible");
-        m.setLayoutProperty("area-flows-outline", "visibility", "visible");
 
         return { destinationPortsIds, legend: null };
       }),
@@ -528,8 +526,6 @@ export const globeViewMachine = createMachine(
             "none"
           );
           m.setLayoutProperty("destination-areas-fill", "visibility", "none");
-          m.setLayoutProperty("area-flows", "visibility", "none");
-          m.setLayoutProperty("area-flows-outline", "visibility", "none");
         }
 
         return {
